face_detection_model/requirements.txt
request==2.31.0
numpy==1.26.2
opencv-python==4.9.0